Refactor example pages: update Cities import path and create shared Cities component
